# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me/profile": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Register new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "User login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid refresh token"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/companies": {
      "get": {
        "operationId": "CompaniesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of companies"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Search companies with filters",
        "tags": [
          "Companies"
        ]
      },
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create company (business role)",
        "tags": [
          "Companies"
        ]
      }
    },
    "/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company details"
          },
          "404": {
            "description": "Company not found"
          }
        },
        "summary": "Get company by ID",
        "tags": [
          "Companies"
        ]
      }
    },
    "/companies/{id}/offers": {
      "post": {
        "operationId": "CompaniesController_addOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Offer added"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Add offer to company",
        "tags": [
          "Companies"
        ]
      }
    },
    "/decide/now": {
      "post": {
        "operationId": "DecideController_decideNow",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Recommended company or null"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Get best option right now",
        "tags": [
          "Decide"
        ]
      }
    },
    "/decide/plan": {
      "post": {
        "operationId": "DecideController_plan",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Plan created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Plan visit for a specific date",
        "tags": [
          "Decide"
        ]
      }
    },
    "/visits": {
      "post": {
        "operationId": "VisitsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Visit created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create new visit",
        "tags": [
          "Visits"
        ]
      }
    },
    "/visits/user/{userId}": {
      "get": {
        "operationId": "VisitsController_findByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of visits"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get visits by user",
        "tags": [
          "Visits"
        ]
      }
    },
    "/offers": {
      "post": {
        "operationId": "OffersController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Offer created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create new offer",
        "tags": [
          "Offers"
        ]
      }
    },
    "/offers/company/{companyId}": {
      "get": {
        "operationId": "OffersController_findByCompany",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of offers"
          },
          "404": {
            "description": "Company not found"
          }
        },
        "summary": "Get offers by company",
        "tags": [
          "Offers"
        ]
      }
    }
  },
  "info": {
    "title": "Decide API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "tags": [],
  "security": [
    {
      "access-token": []
    }
  ],
  "servers": [
    {
      "url": "/api",
      "description": "API"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Liliia",
            "description": "User full name"
          },
          "email": {
            "type": "string",
            "example": "liliia@example.com",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          },
          "city": {
            "type": "string",
            "example": "New York",
            "description": "User city"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "1990-01-01",
            "description": "User date of birth"
          },
          "role": {
            "type": "string",
            "example": "user",
            "description": "User role"
          },
          "acceptPrivacyPolicy": {
            "type": "boolean",
            "example": true,
            "description": "Accept privacy policy"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "city",
          "dateOfBirth",
          "role",
          "acceptPrivacyPolicy"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "GeoDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Point"
          }
        },
        "required": [
          "type"
        ]
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Awesome Cafe"
          },
          "category": {
            "type": "string",
            "example": "cafe"
          },
          "address": {
            "type": "string",
            "example": "123 Main St"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "geo": {
            "example": {
              "type": "Point",
              "coordinates": [
                -74.006,
                40.7128
              ]
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/GeoDto"
              }
            ]
          },
          "workingHours": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "day": {
                  "type": "string"
                },
                "open": {
                  "type": "string"
                },
                "close": {
                  "type": "string"
                }
              }
            }
          },
          "averagePrice": {
            "type": "number",
            "example": 25.5
          }
        },
        "required": [
          "name",
          "category",
          "address",
          "city",
          "geo",
          "workingHours"
        ]
      }
    }
  }
}